requirements_test.txt
coverage==4.3.3
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
django-model-utils>=2.0
-e git+https://github.com/goodmase/django-bitfield.git#egg=django_bitfield
# Additional test requirements go here
